requirements.txt
python-binance==1.0.27
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.0
tensorflow==2.13.0
streamlit==1.26.0
plotly==5.16.1
joblib==1.3.2
python-dotenv==1.0.0
nest-asyncio==1.5.8 